KNN Analyst
K in KNN is the number of nearest neighbours to examine. It is supposed to be specified by the user. So, how can we choose the correct value for K? The general solution is to reserve a part of your data to test the model's accuracy. Then choose k =1, use the training part for modelling, and calculate the accuracy of prediction using all samples in your test set. Repeat this process, increasing the k, and see which k is the best for your model.